+++ protocols/fireface/src/former/ff800.rs
@@ -9,7 +9,7 @@ use super::*;
 #[derive(Default, Debug)]
 pub struct Ff800Protocol;
 
-const MIXER_OFFSET: usize = 0x000080080000;
+const MIXER_OFFSET: u64 = 0x000080080000;
 const OUTPUT_OFFSET: usize = 0x000080081f80;
 const METER_OFFSET: u64 = 0x000080100000;
 const STATUS_OFFSET: u64 = 0x0000801c0000;
@@ -60,7 +60,7 @@ impl RmeFormerOutputOperation for Ff800Protocol {
 }
 
 impl RmeFormerMixerOperation for Ff800Protocol {
-    const MIXER_OFFSET: usize = MIXER_OFFSET;
+    const MIXER_OFFSET: u64 = MIXER_OFFSET;
     const AVAIL_COUNT: usize = 32;
 }
